+++ lib/rubocop/cop/rspec/change_by_zero.rb
@@ -118,7 +118,11 @@ module RuboCop
         # rubocop:enable Metrics/MethodLength
 
         def compound_expectations?(node)
-          %i[and or & |].include?(node.parent.method_name)
+          if node.parent.send_type?
+            %i[and or & |].include?(node.parent.method_name)
+          else
+            node.parent.and_type? || node.parent.or_type?
+          end
         end
 
         def message(change_node)