+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Master (Unreleased)
 
 - Change `RSpec/ContextWording` cop to always report an offense when both `Prefixes` and `AllowedPatterns` are empty. ([@ydah])
+- Fix an error for `RSpec/ChangeByZero` when `change (...) .by (0)` and `change (...)`, concatenated with `and` and `or`. ([@ydah])
 
 ## 3.1.0 (2024-10-01)